pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "seo-agent"
version = "0.1.0"
description = "AI-powered SEO automation tool"
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "dspy-ai>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "tqdm>=4.66.0",
    "requests>=2.30.0",
    "beautifulsoup4>=4.12.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.6",
    "mypy>=1.3.0",
    "pre-commit>=3.3.2",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.12",
]

[project.scripts]
seo-agent = "seo_agent.main:cli"

[tool.ruff]
target-version = "py311"
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = ["E", "F", "I", "W", "D", "N", "C90", "B", "UP", "PL", "RUF"]
ignore = ["D203", "D212"]

# Exclude mock and test files from linting
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "tests/**/*.py",   # Exclude all Python files in tests directory and subdirectories
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["seo_agent"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

# Exclude tests directory from mypy checks
exclude = [
    "tests/*",
    "tests/**/*"
]

# Additional override for tests directory
[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true
